feat: better host restrictions

+++ packages/api/src/modules/paste/paste.service.ts
@@ -1,7 +1,8 @@
 import { EntityRepository } from "@mikro-orm/core";
 import { InjectRepository } from "@mikro-orm/nestjs";
-import { Injectable, Logger, NotFoundException } from "@nestjs/common";
+import { BadRequestException, Injectable, Logger, NotFoundException } from "@nestjs/common";
 import { Cron, CronExpression } from "@nestjs/schedule";
+import { FastifyRequest } from "fastify";
 import { Paste } from "./paste.entity";
 
 @Injectable()
@@ -9,8 +10,12 @@ export class PasteService {
   private log = new Logger("PasteService");
   constructor(@InjectRepository(Paste) private pasteRepo: EntityRepository<Paste>) {}
 
-  async getPaste(pasteId: string) {
+  async getPaste(pasteId: string, request: FastifyRequest) {
     const paste = await this.pasteRepo.findOneOrFail(pasteId);
+    if (!paste.canSendTo(request)) {
+      throw new BadRequestException("Your paste is in another castle.");
+    }
+
     if (paste.expiresAt && paste.expiresAt.getTime() < Date.now()) {
       await this.pasteRepo.removeAndFlush(paste);
       throw new NotFoundException();
